In javascript you can create objects using this syntax:
    var myObject=new Object();
    myObject.add=function(a,b){return a+b};  
    myObject.add(1, 2);

This can be considered a class since it has methods.

The new Object() can also be used within functions:
    function ResultSetData(  ) {
        this.getRow = function( ind ) {
            if ( ind > 0 && ind < this.resultArray.length ) {
                var obj = new Object();
                return obj;
            }
        }
    }

This has both a class, ResultSetData and a method, getRow.
But in this case, the var obj is not a class since it is within
another function.

So in this case, we will exclude these as classes.


git-svn-id: http://ctags.svn.sourceforge.net/svnroot/ctags/trunk@665 c5d04d22-be80-434c-894e-aa346cc9e8e8